moving tests to target framework 4.72
