+++ src/Hive.IO/UnitTests/TestSia2024Record.cs
@@ -1,38 +1,34 @@
 using System.Linq;
 using Hive.IO.Building;
-using NUnit.Framework;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Hive.Tests
+namespace UnitTests
 {
+    [TestClass]
     public class Tests
     {
-        [SetUp]
-        public void Setup()
-        {
-        }
-
-        [Test]
+        [TestMethod]
         public void TestReadRecords()
         {
             var records = Sia2024Record.All();
-            Assert.IsNotEmpty(records);
+            Assert.IsTrue(records.Count() > 0);
             Assert.AreEqual(records.First().RoomType, "1.1 Wohnen Mehrfamilienhaus");
         }
 
-        [Test]
+        [TestMethod]
         public void TestBuildingUseTypes()
         {
             var useTypes = Sia2024Record.BuildingUseTypes();
             Assert.AreEqual(useTypes.Count(), 12);
         }
 
-        [Test]
+        [TestMethod]
         public void TestQualities()
         {
             Assert.AreEqual(Sia2024Record.Qualities().Count(), 3);
         }
 
-        [Test]
+        [TestMethod]
         public void TestQualitiesZielwert()
         {
             var useType = Sia2024Record.BuildingUseTypes().First();
@@ -43,14 +39,14 @@ namespace Hive.Tests
             Assert.AreEqual(record.UValueOpaque, 0.1, 0.0000001);
         }
 
-        [Test]
+        [TestMethod]
         public void TestJson()
         {
             var record = Sia2024Record.All().First();
             var json = record.ToJson();
-            Assert.IsInstanceOf(typeof(string), json);
+            Assert.IsInstanceOfType(json, typeof(string));
             var record2 = Sia2024Record.FromJson(json);
-            Assert.NotNull(record2);
+            Assert.IsNotNull(record2);
         }
     }
 }
